refactor to reduce function size
